Add default response entity
